+++ src/main/java/alec_wam/CrystalMod/items/tools/bat/upgrades/EnderBatUpgrade.java
@@ -1,18 +1,24 @@
 package alec_wam.CrystalMod.items.tools.bat.upgrades;
 
 import java.util.List;
-import java.util.Set;
 
+import alec_wam.CrystalMod.CrystalMod;
+import alec_wam.CrystalMod.api.tools.AttackData;
+import alec_wam.CrystalMod.api.tools.IBatUpgrade;
+import alec_wam.CrystalMod.api.tools.UpgradeData;
+import alec_wam.CrystalMod.items.tools.bat.BatHelper;
+import alec_wam.CrystalMod.items.tools.bat.BatUpgrade;
+import alec_wam.CrystalMod.util.ItemUtil;
+import alec_wam.CrystalMod.util.ModLogger;
+import alec_wam.CrystalMod.util.client.RenderUtil;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.block.model.ItemCameraTransforms.TransformType;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.init.Blocks;
 import net.minecraft.init.Items;
 import net.minecraft.init.SoundEvents;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.EnumParticleTypes;
@@ -21,17 +27,6 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
-import alec_wam.CrystalMod.CrystalMod;
-import alec_wam.CrystalMod.api.tools.AttackData;
-import alec_wam.CrystalMod.api.tools.IBatUpgrade;
-import alec_wam.CrystalMod.api.tools.UpgradeData;
-import alec_wam.CrystalMod.items.tools.bat.BatHelper;
-import alec_wam.CrystalMod.items.tools.bat.BatUpgrade;
-import alec_wam.CrystalMod.util.ChatUtil;
-import alec_wam.CrystalMod.util.ItemUtil;
-import alec_wam.CrystalMod.util.Lang;
-import alec_wam.CrystalMod.util.ModLogger;
-import alec_wam.CrystalMod.util.client.RenderUtil;
 
 public class EnderBatUpgrade extends BatUpgrade {
 
@@ -78,7 +73,7 @@ public class EnderBatUpgrade extends BatUpgrade {
     {
 		if(ent == null)return false;
         double d0 = ent.posX + (ent.getEntityWorld().rand.nextDouble() - 0.5D) * 32.0D;
-        double d1 = ent.posY + (double)(ent.getEntityWorld().rand.nextInt(64) - 32);
+        double d1 = ent.posY + (ent.getEntityWorld().rand.nextInt(64) - 32);
         double d2 = ent.posZ + (ent.getEntityWorld().rand.nextDouble() - 0.5D) * 32.0D;
         return teleportTo(ent, d0, d1, d2);
     }
@@ -138,14 +133,14 @@ public class EnderBatUpgrade extends BatUpgrade {
 
             for (int l = 0; l < short1; ++l)
             {
-                double d6 = (double)l / ((double)short1 - 1.0D);
+                double d6 = l / (short1 - 1.0D);
                 float f = (ent.getEntityWorld().rand.nextFloat() - 0.5F) * 0.2F;
                 float f1 = (ent.getEntityWorld().rand.nextFloat() - 0.5F) * 0.2F;
                 float f2 = (ent.getEntityWorld().rand.nextFloat() - 0.5F) * 0.2F;
-                double d7 = d3 + (ent.posX - d3) * d6 + (ent.getEntityWorld().rand.nextDouble() - 0.5D) * (double)ent.width * 2.0D;
-                double d8 = d4 + (ent.posY - d4) * d6 + ent.getEntityWorld().rand.nextDouble() * (double)ent.height;
-                double d9 = d5 + (ent.posZ - d5) * d6 + (ent.getEntityWorld().rand.nextDouble() - 0.5D) * (double)ent.width * 2.0D;
-                ent.getEntityWorld().spawnParticle(EnumParticleTypes.PORTAL, d7, d8, d9, (double)f, (double)f1, (double)f2);
+                double d7 = d3 + (ent.posX - d3) * d6 + (ent.getEntityWorld().rand.nextDouble() - 0.5D) * ent.width * 2.0D;
+                double d8 = d4 + (ent.posY - d4) * d6 + ent.getEntityWorld().rand.nextDouble() * ent.height;
+                double d9 = d5 + (ent.posZ - d5) * d6 + (ent.getEntityWorld().rand.nextDouble() - 0.5D) * ent.width * 2.0D;
+                ent.getEntityWorld().spawnParticle(EnumParticleTypes.PORTAL, d7, d8, d9, f, f1, f2);
             }
 
             ent.getEntityWorld().playSound(null, d3, d4, d5, SoundEvents.ENTITY_ENDERMEN_TELEPORT, SoundCategory.NEUTRAL, 1.0F, 1.0F);
